+++ src/wp_pbkdf2.c
@@ -259,9 +259,9 @@ static int wp_kdf_pbkdf2_derive(wp_Pbkdf2Ctx* ctx, unsigned char* key,
         int rc;
 
         PRIVATE_KEY_UNLOCK();
-        rc = wc_PBKDF2_ex(key, ctx->password, ctx->passwordSz, ctx->salt,
-            ctx->saltSz, ctx->iterations, keyLen, ctx->mdType, NULL,
-            INVALID_DEVID);
+        rc = wc_PBKDF2_ex(key, ctx->password, (int)ctx->passwordSz, ctx->salt,
+            (int)ctx->saltSz, (int)ctx->iterations, (int)keyLen, ctx->mdType,
+            NULL, INVALID_DEVID);
         PRIVATE_KEY_LOCK();
         if (rc != 0) {
             ok = 0;
@@ -369,9 +369,9 @@ static int wp_kdf_pkcs12_derive(wp_Pbkdf2Ctx* ctx, unsigned char* key,
         int rc;
 
         PRIVATE_KEY_UNLOCK();
-        rc = wc_PKCS12_PBKDF_ex(key, ctx->password, ctx->passwordSz, ctx->salt,
-            ctx->saltSz, ctx->iterations, keyLen, ctx->mdType, ctx->keyUse,
-            NULL);
+        rc = wc_PKCS12_PBKDF_ex(key, ctx->password, (int)ctx->passwordSz,
+            ctx->salt, (int)ctx->saltSz, (int)ctx->iterations, (int)keyLen,
+            ctx->mdType, ctx->keyUse, NULL);
         PRIVATE_KEY_LOCK();
         if (rc != 0) {
             ok = 0;
